+++ types/resources/ItemPrice.d.ts
@@ -91,16 +91,6 @@ declare module 'chargebee' {
         input?: FindApplicableItemPricesInputParam,
         headers?: ChargebeeRequestHeader,
       ): Promise<ChargebeeResponse<FindApplicableItemPricesResponse>>;
-
-      /**
-       * @deprecated This method is deprecated and will be removed in a future version.
-       */
-
-      moveItemPrice(
-        item_price_id: string,
-        input: MoveItemPriceInputParam,
-        headers?: ChargebeeRequestHeader,
-      ): Promise<ChargebeeResponse<MoveItemPriceResponse>>;
     }
 
     export interface CreateResponse {
@@ -132,10 +122,6 @@ declare module 'chargebee' {
     export interface FindApplicableItemPricesResponse {
       list: { item_price: ItemPrice }[];
       next_offset?: string;
-    }
-
-    export interface MoveItemPriceResponse {
-      item_price: ItemPrice;
     }
 
     export interface Tier {
@@ -282,10 +268,6 @@ declare module 'chargebee' {
       item_id?: string;
       'sort_by[asc]'?: string;
       'sort_by[desc]'?: string;
-    }
-    export interface MoveItemPriceInputParam {
-      destination_item_id: string;
-      variant_id?: string;
     }
     export interface TaxDetailCreateInputParam {
       tax_profile_id?: string;